+++ cipher_reactivity/setup.cfg
@@ -1,0 +1,33 @@
+[metadata]
+name = cipher
+version = 0.0.1a
+long_description = file: README.md
+long_description_content_type = text/markdown
+
+[options]
+include_package_data = True
+packages = find:
+# package_dir =
+# 	=src
+python_requires = >=3.7
+install_requires = 
+	graphviz
+	mongoengine
+	pymongo
+	requests
+	tomli
+
+# [options.packages.find]
+# include =
+# 	src
+
+[options.package_data]
+cipher_reactivity = data/*
+
+[options.extras_require]
+dev =
+	black
+	flake8
+	pytest	
+test =
+	pytest
